requirements.txt
requests
supabase
python-dotenv
fastapi
uvicorn[standard]
africastalking
streamlit
streamlit-folium
pandas